+++ src/test/java/cvds/todo/backend/service/TaskServiceTest.java
@@ -1,0 +1,4 @@
+package cvds.todo.backend.service;
+
+public class TaskServiceTest {
+}
